+++ netcfgbu/jumphosts.py
@@ -1,8 +1,14 @@
+"""
+This module contains the code providing jump host feature functionality
+so that any device in inventory that requires a proxy server can use
+the netcfgbu tool.
+"""
+
 # -----------------------------------------------------------------------------
 # System Imports
 # -----------------------------------------------------------------------------
 
-from typing import Optional, List, AnyStr
+from typing import Optional, List, Dict, AnyStr
 import asyncio
 from urllib.parse import urlparse
 
@@ -23,18 +29,76 @@ from .logger import get_logger
 
 
 class JumpHost(object):
-    active_jumphosts = list()
+    """
+    A JumpHost instance is used to provide a tunnel connection so that any
+    device in the inventory that requires one can use it.
+    """
 
-    def __init__(self, spec: JumphostSpec):
+    available = list()
+
+    def __init__(self, spec: JumphostSpec, field_names: List[AnyStr]):
+        """
+        Prepare a jump host instance for potential use.  This method
+        does not connect to the proxy system.
+        Parameters
+        ----------
+        spec:
+            The jumphost configuraiton
+
+        field_names:
+            List of inventory field names that are used to prepare any
+            necessary filtering functionality
+        """
         self._spec = spec
         self.filters = list()
         self._conn = None
+        self._init_filters(field_names)
 
     @property
     def tunnel(self):
+        """
+        Returns the SSH client connection of the jump-host for use as `tunnel` when
+        connecting to a target device.  If the SSH client does not exist, then raise
+        a RuntimeError.
+        """
+        if not self.is_active:
+            raise RuntimeError(
+                f"Attempting to use JumpHost {self.name}, but not connected"
+            )
         return self._conn
 
+    @property
+    def name(self):
+        """ Returns the string-name of the jump host"""
+        return self._spec.name
+
+    @property
+    def is_active(self):
+        """ Return True if the jumphost is connected, False otherwise """
+        return bool(self._conn)
+
+    def _init_filters(self, field_names):
+        """ Called only by init, prepares the jump host filter functions to later use """
+        include, exclude = self._spec.include, self._spec.exclude
+        if include:
+            self.filters.append(
+                create_filter(
+                    constraints=include, field_names=field_names, include=True
+                )
+            )
+
+        if exclude:
+            self.filters.append(
+                create_filter(
+                    constraints=exclude, field_names=field_names, include=False
+                )
+            )
+
     async def connect(self):
+        """
+        Connects to the jumphost system so that it can be used later as the
+        tunnel to connect to other devices.
+        """
         proxy_parts = urlparse("ssh://" + self._spec.proxy)
 
         conn_args = dict(host=proxy_parts.hostname, known_hosts=None)
@@ -45,12 +109,16 @@ class JumpHost(object):
             conn_args["port"] = proxy_parts.port
 
         async def connect_to_jh():
+            """ obtain the SSH client connection """
             self._conn = await asyncssh.connect(**conn_args)
 
         await asyncio.wait_for(connect_to_jh(), timeout=self._spec.timeout)
-        self.__class__.active_jumphosts.append(self)
 
     def filter(self, inv_rec):
+        """
+        This function returns True if this jump host is required to support the given
+        inventory record.  Returns False otherwise.
+        """
         return any(_f(inv_rec) for _f in self.filters)
 
 
@@ -61,41 +129,62 @@ class JumpHost(object):
 # -----------------------------------------------------------------------------
 
 
-async def init_jumphosts(jumphosts: List[JumphostSpec], field_names=List[str]):
-    try:
-        for jh_spec in jumphosts:
-            await create_jumphost_client(jh_spec, field_names)
-    except asyncssh.Error:
-        raise RuntimeError("Failed to connect to jumphost(s)")
+def init_jumphosts(jumphost_specs: List[JumphostSpec], inventory: List[Dict]):
+    """
+    Initialize the required set of Jump Host instances so that they can be used
+    when netcfgbu attempts to access devices that require the use of jump
+    hosts.
+
+    Parameters
+    ----------
+    jumphost_specs:
+        List of jump host specs from the app config instance
+
+    inventory:
+        List of inventory records; these are used to determine which, if any,
+        of the configured jump hosts are actually required for use given any
+        provided inventory filtering.
+    """
+    field_names = inventory[0].keys()
+
+    # create a list of jump host instances so that we can determine which, if
+    # any, will be used during the execution of the command.
+
+    jh_list = [JumpHost(spec, field_names=field_names) for spec in jumphost_specs]
+
+    req_jh = {
+        use_jh
+        for rec in inventory
+        if (use_jh := first(jh for jh in jh_list if jh.filter(rec)))
+    }
+
+    JumpHost.available = list(req_jh)
 
 
-async def create_jumphost_client(
-    jh_spec: JumphostSpec, field_names: List[AnyStr]
-) -> asyncssh.SSHClientConnection:
+async def connect_jumphosts():
+    """
+    This coroutine is used to connect to all of the required jump host servers.  This
+    should be called before attempting to run any of the SSH device tasks, such as
+    login or backup.
 
+    Returns
+    -------
+    True if all required jump host servers are connected.
+    False otherwise; check log errors for details.
+    """
     log = get_logger()
-    jh = JumpHost(jh_spec)
+    ok = True
 
-    try:
-        await jh.connect()
-        log.info(f"JUMPHOST: connected to {jh_spec.name}")
-    except asyncssh.Error as exc:
-        log.error(f"JUMPHOST: connect to {jh_spec.name} failed: {str(exc)}")
-        raise
+    for jh in JumpHost.available:
+        try:
+            await jh.connect()
+            log.info(f"JUMPHOST: connected to {jh.name}")
 
-    if jh_spec.include:
-        jh.filters.append(
-            create_filter(
-                constraints=jh_spec.include, field_names=field_names, include=True
-            )
-        )
+        except asyncssh.Error as exc:
+            log.error(f"JUMPHOST: connect to {jh.name} failed: {str(exc)}")
+            ok = False
 
-    if jh_spec.exclude:
-        jh.filters.append(
-            create_filter(
-                constraints=jh_spec.exclude, field_names=field_names, include=False
-            )
-        )
+    return ok
 
 
 def get_jumphost(inv_rec: dict) -> Optional[JumpHost]:
@@ -104,4 +193,4 @@ def get_jumphost(inv_rec: dict) -> Optional[JumpHost]:
     for the given inventory record.  If this record does not require the
     use of a jumphost, then return None.
     """
-    return first(jh for jh in JumpHost.active_jumphosts if jh.filter(inv_rec))
+    return first(jh for jh in JumpHost.available if jh.filter(inv_rec))